fix(garden): keep seed reader effects stable
