+++ app/src/components/SeedDocumentView.tsx
@@ -2,17 +2,12 @@ import { useId, type Ref } from 'react';
 import type {
   Seed,
   SeedArtifactReference,
-  SeedNote,
 } from '../types/generated';
 import { Markdown } from './Markdown';
 import { MarkdownReader, type MarkdownAnnotationsSendHandle } from './MarkdownReader';
 import { seedMarkdownSource } from './MarkdownReader/documentSource';
+import { currentSeedArtifacts, type SeedDocumentNote } from './seedArtifacts';
 import './SeedDocumentView.css';
-
-/** SeedNote is deliberately forward-compatible with the later artifact verbs. */
-export type SeedDocumentNote = Omit<SeedNote, 'artifact'> & {
-  artifact?: SeedArtifactReference;
-};
 
 /** The one read model shared by the panel drill and the docked seed tile. */
 export interface SeedDocument {
@@ -40,22 +35,6 @@ function artifactKey(artifact: SeedArtifactReference): string {
     artifact.repository ?? '',
     artifact.url ?? '',
   ].join('\0');
-}
-
-/** Project the current set by replaying the newest-first log chronologically. */
-export function currentSeedArtifacts(notes: readonly SeedDocumentNote[]): SeedArtifactReference[] {
-  const current = new Map<string, SeedArtifactReference>();
-  for (let i = notes.length - 1; i >= 0; i -= 1) {
-    const note = notes[i];
-    if (!note.artifact) continue;
-    const key = artifactKey(note.artifact);
-    if (note.kind === 'attach') {
-      current.set(key, note.artifact);
-    } else if (note.kind === 'detach') {
-      current.delete(key);
-    }
-  }
-  return Array.from(current.values());
 }
 
 function formatTimestamp(iso: string): string {
